Use a more precise definition of state
ALE's state serialization includes too many things that are not really part of
state in the MDP or game-tree sense (like frame number). For games not using
paddles, getRAM() is what we want. But for games using paddles, we need the
current paddle position too. As a result, use a tuple (RAM, paddle_pos) for
state. This requires a modified ALE, since paddle position isn't exposed by
default.