Reset the game at start and after each ID iteration
A reset at the start might not be strictly necessary, but there is some
weirdness around whether resetting produces a bit-identical state to a freshly
loaded ROM in all cases, so just reset to be safe.

Resetting after each ID iteration is needed, because even though we've "backed
out" the actions by recursively restoring state, the RAM of a restored state is
stale until a new act() is taken. We previously tried to address this by patching
ALE, but Jesse Farebrother upstream pointed out some issues there, and this is
a simpler fix anyway.