Add a special-case fix for 'trondead'
The game 'trondead' starts in a 'dead' initial state: we need to take one dummy
action to get the emulation going properly it seems.

The game 'klax' also starts in a dead initial state, but does not seem to have
as easy a fix. The other 102 ALE-supported ROMs behave as expected in this
regard.