Updated 10k validation results for camera ready
Re-ran with a slight tweak to the DFS used in iterative deepening (see previous
commit), to re-expand a previously seen state *if* we newly encounter it in
a more shallow part of the tree. This reduces the number of games where BFS and
ID disagree from 14 to 10 (out of 103 tested).